+++ CHANGELOG.md
@@ -59,6 +59,7 @@ The following lists show the changes to the library grouped by domain.
 #### Various
 
 * fixing [`ally.maintain.disabled`][ally/maintain/disabled] to properly handle `tabindex` attribute changes
+* fixing [`ally.maintain.disabled`][ally/maintain/disabled] to properly disengage within ShadowHosts - [issue #107](https://github.com/medialize/ally.js/issues/107), [PR #108](https://github.com/medialize/ally.js/pull/108)
 
 #### Internals